requirements.txt
opencv-python==4.12.0.88
numpy==1.24.3
Pillow==10.0.0
matplotlib==3.7.2
scipy==1.16.3
scikit-image==0.25.2
python-magic-bin==0.4.14  # Windows用户使用这个